Optimize and de-clutter
I'd ideally like to sort (reverse) with `awk`, and it's doable, but
unfortunately I can't remember how to do it, and my search online isn't
yielding any results.

It's a real shame to have to re-execute `awk` each line, so I will
probably revisit this block, because I feel like it could certainly be
further optimized. This is still, however, absolutely an improvement.

If there are any issues with this block, please let me know, so that I
might hopefully address them.